feat(ZNTA-2380): save user permissions in header data reducer
